+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.1.4
 *   Feature : Added number and CVV obfuscation
+*   Feature : Added input decorations
 
 ## 0.1.3
 *   Feature : Applied Default ThemeData.